todo_drf_auth_main/requirements.txt
asgiref==3.5.1
backports.zoneinfo==0.2.1
black==22.3.0
cffi==1.15.0
click==8.1.3
cryptography==37.0.2
Django==4.0.4
django-cors-headers==3.11.0
django-filter==21.1
djangorestframework==3.13.1
flake8==4.0.1
importlib-metadata==4.11.3
invoke==1.7.0
Markdown==3.3.6
mccabe==0.6.1
mypy==0.950
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
platformdirs==2.5.2
psycopg2-binary==2.8.5
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
PyJWT==1.7.1
pyparsing==3.0.8
python-dotenv==0.20.0
pytz==2022.1
sqlparse==0.4.2
tomli==2.0.1
typing_extensions==4.2.0
zipp==3.8.0
